adds renderer component; moves line-drawing methods to it
